[Architecture] Fix line-limit guardrails on PR #17 | No playtest needed (CI-only)
The ArchitectureGuardrailTests.RuntimeFilesStayUnderLineLimitOrDoNotGrowPastBaseline
test was failing because two runtime files exceeded their limits:

- NetworkLootItem.cs: 704 lines vs baseline 703 (RoundManager.Instance ->
  RoundManagerRegistry.Current rewrite split a one-liner into two). Collapsed
  back to a single line via 'is { } activeRound' pattern.
- PrototypeNetworkBootstrapper.cs: 523 lines vs default 400 (added 159 lines
  for ship-scene flow). Split into a partial class:
  - Main file (342 lines): lifecycle, NetworkManager wiring, RoundManager
    spawn, ship/planet flow orchestration.
  - PrototypeNetworkBootstrapper.Spawning.cs (194 lines): SpawnLoot,
    SpawnMonsters, ship-part placement, and DisableLaunchpadInterference.

All SerializeFields kept in the main partial to avoid any risk of disturbing
serialized references in committed prefabs/scenes.
