+++ Space Game/Assets/Scripts/Networking/PrototypeNetworkBootstrapper.Spawning.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: abe90948bc3f4cbc9a8b0f576e5bac7c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
